Fix mapping of PDButton
Nim appears to already map enum values to a bitfield
by doing a shl internally. Doing the shl in the binding code was
forcing a "double encoding" problem. This was causing the `PDButtons`
set to not actually work like a real nim `set`.

This change also removes the 'check' function, since we
can now use regular Nim set operations. For example, we
could just do:

  kButtonA in buttons